requirements.txt
discord.py==2.2.2
openai==0.27.0
python-dotenv==1.0.0
googletrans==4.0.0rc1